+++ cart/package.json
@@ -55,6 +55,7 @@
     "@types/jest": "^29.2.3",
     "@types/lodash": "^4.14.186",
     "@types/react-table": "^7.7.14",
-    "@types/uuid": "^8.3.4"
+    "@types/uuid": "^8.3.4",
+    "dotenv": "^16.1.4"
   }
 }
