+++ cart/src/components/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Nav from './Nav';
+import Nav from './shared/Nav';
 import './App.css';
  
 function App() {
