+++ cart/src/components/admin/AdminProductList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Product} from "../types/types"
+import {Product} from "../../types/types"
 import AdminProductPlank from "./AdminProductPlank"
 import './Admin.css';
 
@@ -9,13 +9,7 @@ interface Props {
 
 const  AdminProductList:React.FC <Props> = ({productList}) => {
 
-
-  const handleAdd = (product: Product) => {
-      console.log("adding "+product.product_id)
-  }
-
   const renderProductlist = ():JSX.Element[]  => {
-
     return productList.map(item => {
       return (
         <div key={item.product_id} className="prod-list-item" >
