+++ cart/src/components/admin/ProductForm.tsx
@@ -1,5 +1,5 @@
 import React, {useState} from 'react';
-import {Product} from "../types/types"
+import {Product} from "../../types/types"
 import './Admin.css';
 
 interface Props {
