+++ cart/src/components/cart/BasketList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {OrderItem} from "../types/types"
+import {OrderItem} from "../../types/types"
 import BasketPlank from "./BasketPlank"
 import './Cart.css';
