+++ cart/src/components/cart/ShoppingProductList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {Product} from "../types/types"
+import {Product} from "../../types/types"
 import ShoppingProductPlank from "./ShoppingProductPlank"
 import './Cart.css';
