+++ cart/src/components/cart/ShoppingProductPlank.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import {Product} from "../types/types"
-import {toDollarString} from "../utils"
+import {Product} from "../../types/types"
+import {toDollarString} from "../../utils"
 import './Shopping.css';
